edge-agent/requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ONNX Runtime (optional - for ML models)
# onnxruntime>=1.16.0

# Monitoring
python-json-logger>=2.0.0
prometheus-client>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
